add background

+++ src/components/card/Card.js
@@ -5,7 +5,7 @@ import './Card.css';
 const CardItem = ({day}) => {
 
   const ms = day.dt * 1000;
-  const dayName = new Date(ms).toLocaleDateString('ru', {weekday: 'long', day: 'numeric', month: 'long'});
+  const dayName = new Date(ms).toLocaleDateString('ru', {weekday: 'short', day: 'numeric', month: 'long'});
   const imgURL = "owf owf-" + day.weather[0].id + " owf-5x";
 
   return (
